Final-project-inf-119/requirements.txt
# Core dependencies
google-generativeai>=0.8.0
gradio>=5.0.0
langchain>=0.1.0
langchain-google-genai>=0.0.6

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
pydantic>=2.0.0

# Verb conjugation libraries (for generated code)
mlconjug3>=3.8.0
verbecc>=1.6.0
